+++ benches/benchmark.rs
@@ -31,6 +31,13 @@ fn pass_bench(_dummy: usize) -> Wave {
     )
 }
 
+#[allow(clippy::precedence)]
+fn netpass_bench(_dummy: usize) -> Wave {
+    let x = Net::wrap(Box::new(dc((1.0, 2.0))));
+    let y = Net::wrap(Box::new(pass()));
+    Wave::render(44100.0, 1.0, &mut (x * 2.0 >> pass() + y >> pass()))
+}
+
 fn wavetable_bench(_dummy: usize) -> Wave {
     Wave::render(44100.0, 1.0, &mut (saw_hz(110.0)))
 }
@@ -81,6 +88,7 @@ fn phaser_bench(_dummy: usize) -> Wave {
 }
 
 fn criterion_benchmark(c: &mut Criterion) {
+    c.bench_function("netpass", |b| b.iter(|| netpass_bench(black_box(0))));
     c.bench_function("sine", |b| b.iter(|| sine_bench(black_box(0))));
     c.bench_function("resynth", |b| b.iter(|| resynth_bench(black_box(0))));
     c.bench_function("pass", |b| b.iter(|| pass_bench(black_box(0))));