+++ src/hacker.rs
@@ -345,6 +345,13 @@ pub fn sine_hz(f: f32) -> An<Pipe<Constant<U1>, Sine>> {
     constant(f) >> sine()
 }
 
+/// Sine oscillator with initial `phase` in 0...1.
+/// - Input 0: frequency (Hz)
+/// - Output 0: sine wave
+pub fn sine_phase(phase: f32) -> An<Sine> {
+    An(Sine::with_phase(phase))
+}
+
 /// Rossler dynamical system oscillator.
 /// - Input 0: frequency. The Rossler oscillator exhibits peaks at multiples of this frequency.
 /// - Output 0: system output
@@ -822,11 +829,8 @@ pub fn hold_hz(f: f32, variability: f32) -> An<Pipe<Stack<Pass, Constant<U1>>, H
 /// use fundsp::hacker::*;
 /// fir(Frame::from([0.5, 1.0, 0.5]));
 /// ```
-pub fn fir<X: ConstantFrame<Sample = f32>>(weights: X) -> An<Fir<X::Size>>
-where
-    X::Size: Size<f32>,
-{
-    An(Fir::new(weights.frame()))
+pub fn fir<X: ConstantFrame<Sample = f32>>(weights: X) -> An<Fir<X::Size>> {
+    An(Fir::new(weights))
 }
 
 /// Create a 3-point symmetric FIR from desired `gain` (`gain` >= 0) at the Nyquist frequency.
@@ -863,8 +867,8 @@ pub fn multitick<N: Size<f32>>() -> An<Tick<N>> {
     An(Tick::new())
 }
 
-/// Fixed delay of `t` seconds.
-/// Delay time is rounded to the nearest sample. The minimum delay is one sample.
+/// Fixed delay of `t` seconds (`t` >= 0).
+/// Delay time is rounded to the nearest sample. The minimum delay is zero samples.
 /// - Allocates: the delay line.
 /// - Input 0: signal.
 /// - Output 0: delayed signal.
@@ -879,7 +883,8 @@ pub fn delay(t: f32) -> An<Delay> {
 }
 
 /// Tapped delay line with cubic interpolation.
-/// Minimum and maximum delay times are in seconds.
+/// Minimum and maximum delay times are in seconds (`min_delay`, `max_delay` >= 0).
+/// The minimum possible delay is one sample.
 /// - Allocates: the delay line.
 /// - Input 0: signal.
 /// - Input 1: delay time in seconds.
@@ -896,7 +901,8 @@ pub fn tap(min_delay: f32, max_delay: f32) -> An<Tap<U1>> {
 
 /// Tapped delay line with cubic interpolation.
 /// The number of taps is `N`.
-/// Minimum and maximum delay times are in seconds.
+/// Minimum and maximum delay times are in seconds (`min_delay`, `max_delay` >= 0).
+/// The minimum possible delay is one sample.
 /// - Allocates: the delay line.
 /// - Input 0: signal.
 /// - Inputs 1...N: delay time in seconds.